// src/app/models/texts.model.ts
export class Texts {
    text: string = "Sample Text";
    fontFamily: string = "Arial";
    fontSize: number = 50;
    textColor: string = "#FFFFFF";
    bold: boolean = false;
    italic: boolean = false;
    strock: boolean = false;
    shadowColor: string = "#FFFFFF";
    shadowX: number = 0;
    shadowY: number = 0;
    top: number = 180;
    left: number = 300;
    strockWidth: number = 1;
    lineheight: number = 100;
    shadowBlur: number = 4;
    fillColor: string = "#f16501";
  }
  